v5.x-scala-js: move ArithUtils to sdk package
