+++ sdk/shared/src/main/scala/org/ergoplatform/sdk/utils/ArithUtils.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.utils
+package org.ergoplatform.sdk.utils
 
 object ArithUtils {
